--- SCRATCH
@@ -1,33 +1,0 @@
-[DEFAULT]
-Latency = 450
-JitterDistribution = Pareto
-JitterMean = 50
-JitterVariance = 30
-Correlation = 35
-LossModel = Bernoulli
-p = 0.00000013
-h = 0
-1-h = 0
-1-k = 0
-CorruptChance = 0.0001
-CorruptCorrelation = 5
-DuplicateChance = 0.0001
-DuplicateCorrelation = 10
-Reorder = yes
-
-[Aggressive]
-Latency = 450
-JitterDistribution = Pareto
-JitterMean = 50
-JitterVariance = 30
-Correlation = 35
-LossModel = gemodel
-p = 0.000196854
-h = 0.0109547
-1-h = 0.436487
-1-k = 0
-CorruptChance = 0.0001
-CorruptCorrelation = 5
-DuplicateChance = 0.0001
-DuplicateCorrelation = 10
-Reorder = yes
