+++ config.ini
@@ -12,18 +12,14 @@ DuplicateCorrelation = 0
 Reorder = False
 
 [Aggressive]
-RateBurst = 32k
-RateLatencyDrop = 200
-PeakRate = 1m
-MinBurst = 1520
-LatencyDistribution = Normal
-LatencyMean = 150
-LatencyVariance = 70
-LatencyCorrelation = 22
-DropProbability = 0.00001
-DropCorrelation = 15
-CorruptProbability = 0.00001
-CorruptCorrelation = 25
-DuplicateChance = 0.000001
-DuplicateCorrelation = 5
-Reorder = True
+LatencyDistribution = normal
+LatencyMean = 40
+LatencyVariance = 5
+LatencyCorrelation = 5
+DropProbability = 3
+DropCorrelation = 25
+CorruptProbability = 0.01
+CorruptCorrelation = 5
+DuplicateChance = 0.001
+DuplicateCorrelation = 0
+Reorder = False
