--- setup.cfg
@@ -1,4 +1,0 @@
-[egg_info]
-tag_build = 
-tag_date = 0
-
