+++ tc/__init__.py
@@ -86,25 +86,6 @@ class DegradeInvoker:
 
     def __init__(self, command: Command):
         self._command = command
-        self._command_list = []  # type :List[Command]
-
-    def set_command(self, command: Command):
-        self.command = command
-
-    def get_command(self):
-        print(self.command.__class__.__name__)
-
-    def add_command_to_list(self, command: Command):
-        self._command_list.append(command)
-
-    def execute_commands(self):
-        """
-        Execute all the saved commands, then empty the list.
-        """
-        for cmd in self._command_list:
-            cmd.execute()
-
-        self._command_list.clear()
 
     def invoke(self):
         self._command.execute()
