--- test/test_aggressive.py
@@ -1,24 +1,0 @@
-from unittest import TestCase
-
-
-class TestAggressive(TestCase):
-    def test_reorder(self):
-        self.fail()
-
-    def test_duplicate(self):
-        self.fail()
-
-    def test_make_command(self):
-        self.fail()
-
-    def test_corrupt(self):
-        self.fail()
-
-    def test_latency(self):
-        self.fail()
-
-    def test_drop(self):
-        self.fail()
-
-    def test_rate(self):
-        self.fail()
